requirements.txt
pandas
numpy
requests
aiohttp
python-jose
pyperclip
pyyaml